+++ src/services/artist.service.js
@@ -1,23 +1,31 @@
-const Artist = require('../../models/Artist');
+const Artist = require("../../models/Artist");
+const Album = require("../../models/Album");
+const Song = require("../../models/Song");
+const AppError = require("../utils/appError");
 
-const createArtist = async (artistData) => {
-    const newArtist = await Artist.create(artistData);
-    return newArtist;
-};
-
-const getAllArtist = async () => {
-    return await Artist.find();
-};
-
-const updateArtist = async (id, updates) => {
-    return await Artist.findByIdAndUpdate(id, updates, {
+const clean = (data) =>
+    Object.fromEntries(
+        Object.entries(data).filter(([, value]) => value !== undefined),
+    );
+const createArtist = (artistData) => Artist.create(clean(artistData));
+const getAllArtist = () => Artist.find().lean();
+const updateArtist = (id, updates) =>
+    Artist.findByIdAndUpdate(id, clean(updates), {
         runValidators: true,
-        new: true
+        new: true,
     });
-};
 
 const deleteArtist = async (id) => {
-    return await Artist.findByIdAndDelete(id);
+    const [albumCount, songCount] = await Promise.all([
+        Album.countDocuments({ artist: id }),
+        Song.countDocuments({ artist: id }),
+    ]);
+    if (albumCount || songCount)
+        throw new AppError(
+            "Artist is referenced by existing albums or songs",
+            409,
+        );
+    return Artist.findByIdAndDelete(id);
 };
 
 module.exports = { createArtist, getAllArtist, updateArtist, deleteArtist };
